高抬腿数据提交成功

+++ app/src/main/java/com/fjp/skeletalmuscle/app/weight/ExerciseIntensityLayout.kt
@@ -30,16 +30,16 @@ class ExerciseIntensityLayout(context: Context, attrs: AttributeSet) : LinearLay
 
     }
 
-    fun setValue(warmupTime: Float, fatBurningTime: Float, cardioTime: Float, breakTime: Float) {
+    fun setValue(warmupTime: Int, fatBurningTime: Int, cardioTime: Int, breakTime: Int) {
         val warmupTimeLP = LinearLayoutCompat.LayoutParams(0,  // 宽度设置为包含布局的权重比例
-            LinearLayoutCompat.LayoutParams.WRAP_CONTENT, warmupTime)
+            LinearLayoutCompat.LayoutParams.WRAP_CONTENT, warmupTime.toFloat())
         warmupTimeTv.layoutParams = warmupTimeLP
 
-        val fatBurningTimeLP = LinearLayoutCompat.LayoutParams(0, LinearLayoutCompat.LayoutParams.WRAP_CONTENT, fatBurningTime)
+        val fatBurningTimeLP = LinearLayoutCompat.LayoutParams(0, LinearLayoutCompat.LayoutParams.WRAP_CONTENT, fatBurningTime.toFloat())
         fatBurningTimeTv.layoutParams = fatBurningTimeLP
-        val cardioTimeLP = LinearLayoutCompat.LayoutParams(0, LinearLayoutCompat.LayoutParams.WRAP_CONTENT, cardioTime)
+        val cardioTimeLP = LinearLayoutCompat.LayoutParams(0, LinearLayoutCompat.LayoutParams.WRAP_CONTENT, cardioTime.toFloat())
         cardioTimeTv.layoutParams = cardioTimeLP
-        val breakTimeLP = LinearLayoutCompat.LayoutParams(0, LinearLayoutCompat.LayoutParams.WRAP_CONTENT, breakTime)
+        val breakTimeLP = LinearLayoutCompat.LayoutParams(0, LinearLayoutCompat.LayoutParams.WRAP_CONTENT, breakTime.toFloat())
         breakTimeTv.layoutParams = breakTimeLP
     }
 }
